requirements.txt
fastapi
uvicorn
httpx
pydantic>=2.0
openai>=1.0.0
python-dotenv
llama-cpp-python